Update and rename GiveFactionAbility.js to GiveSuperpowers.js
